Split country_telephone_data.js into index.js & data.js
