+++ package.json
@@ -2,7 +2,7 @@
     "name": "country-telephone-data",
     "version": "0.6.0",
     "description": "Has country codes for all countries in the world. Please send a PR if you see something missing or find a mistake.",
-    "main": "country_telephone_data.js",
+    "main": "index.js",
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1"
     },
